+++ src/Rpc/RpcServer.cpp
@@ -563,12 +563,12 @@ bool RpcServer::f_on_blocks_list_json(const F_COMMAND_RPC_GET_BLOCKS_LIST::reque
   }
 
   uint32_t print_blocks_count = 30;
-  uint32_t last_height = req.height - print_blocks_count;
+  uint32_t last_height = static_cast<uint32_t>(req.height - print_blocks_count);
   if (req.height <= print_blocks_count)  {
     last_height = 0;
   }
 
-  for (uint32_t i = req.height; i >= last_height; i--) {
+  for (uint32_t i = static_cast<uint32_t>(req.height); i >= last_height; i--) {
     Hash block_hash = m_core.getBlockHashByIndex(static_cast<uint32_t>(i));
     if (!m_core.hasBlock(block_hash)) {
       throw JsonRpc::JsonRpcError{
@@ -992,7 +992,7 @@ void RpcServer::fill_block_header_response(const BlockTemplate& blk, bool orphan
 	response.difficulty = m_core.getBlockDifficulty(index);
 	response.reward = get_block_reward(blk);
 	BlockDetails blkDetails = m_core.getBlockDetails(hash);
-	response.num_txes = blkDetails.transactions.size();
+	response.num_txes = static_cast<uint32_t>(blkDetails.transactions.size());
 	response.block_size = blkDetails.blockSize;
 }
 
@@ -1052,7 +1052,7 @@ bool RpcServer::on_get_block_headers_range(const COMMAND_RPC_GET_BLOCK_HEADERS_R
 		return false;
 	}
 
-	for (uint64_t h = req.start_height; h <= req.end_height; ++h) {
+	for (uint32_t h = static_cast<uint32_t>(req.start_height); h <= static_cast<uint32_t>(req.end_height); ++h) {
 		Crypto::Hash block_hash = m_core.getBlockHashByIndex(h);
 		CryptoNote::BlockTemplate blk = m_core.getBlockByHash(block_hash);
